chore: move current metrics sink to metrics folder
