+++ src/sinks/greptimedb/metrics/batch.rs
