+++ src/sinks/greptimedb/metrics/integration_tests.rs
